Putting everything in the folder
