+++ Python/smartList.txt
