+++ Python/tempAdvanced.py
