+++ Python/tempSmart.py
